+++ ChangeLog.md
@@ -4,8 +4,8 @@
 
 * Add the `RunWithable` class.
 
-* Add the `within` and `liftActionIO` functions for use with
-  `Withable`.
+* Add the `within`, `liftActionIO` and `liftThrow` functions for use
+  with `Withable`.
 
 ## 0.2.0.0 -- 2017-07-07